if you're gonna use doubles internally, use doubles *consistently* internally
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17150 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/ftgl/src/FTPoint.cpp
@@ -43,10 +43,10 @@ bool operator != (const FTPoint &a, const FTPoint &b)
 
 FTPoint FTPoint::Normalise()
 {
-    float norm = sqrt(values[0] * values[0]
+    double norm = sqrt(values[0] * values[0]
                        + values[1] * values[1]
                        + values[2] * values[2]);
-    if(norm == 0.f)
+    if(norm == 0.0)
     {
         return *this;
     }
